Rename SCRUM-84: 타입 파일 이름에 types 추가
